+++ src/wii/init.c
@@ -153,7 +153,7 @@ static void armbootnow(void) {
 		panic("ARMBootNow: Couldn't find LaunchIOS trampoline pointer");
 
 	/* write the pointer to our code there instead */
-	sram[trampoline_off] = (u32)virtToPhys(armbuf) + armbuf[0];
+	sram[trampoline_off] = (uintptr_t)virtToPhys(armbuf) + armbuf[0];
 	log_printf("set trampoline ptr to %08x\r\n", sram[trampoline_off]);
 
 	/* tell MINI not to reset us pretty please */
